fix: bugs in other tests so all pass
