+++ packages/gossipsub/test/partial-messages/utils.ts
@@ -1,18 +1,6 @@
 import { stop } from '@libp2p/interface'
-import pWaitFor from 'p-wait-for'
 import { createComponents, connectPubsubNodes } from '../utils/create-pubsub.js'
 import type { GossipSubAndComponents } from '../utils/create-pubsub.js'
-
-export async function waitForStreamsReady (a: GossipSubAndComponents, b: GossipSubAndComponents): Promise<void> {
-  await pWaitFor(() => {
-    const gsA = a.pubsub as any
-    const gsB = b.pubsub as any
-    const bId = b.components.peerId.toString()
-    const aId = a.components.peerId.toString()
-    return gsA.peers.has(bId) && gsB.peers.has(aId) &&
-      gsA.streamsOutbound.has(bId) && gsB.streamsOutbound.has(aId)
-  }, { timeout: 10000 })
-}
 
 export interface TwoNodeContext {
   nodeA: GossipSubAndComponents
@@ -31,7 +19,6 @@ export async function setupTwoNodes (): Promise<TwoNodeContext> {
     }
   })
   await connectPubsubNodes(nodeA, nodeB)
-  await waitForStreamsReady(nodeA, nodeB)
   return { nodeA, nodeB }
 }
