+++ packages/utils/src/mock-muxer.ts
@@ -271,6 +271,7 @@ class MockMuxer extends AbstractStreamMuxer<MockMuxedStream> {
     this.log.trace('createStream %s %s', direction, id)
 
     return new MockMuxedStream({
+      protocol: '',
       ...this.streamOptions,
       ...options,
       id,
@@ -278,8 +279,7 @@ class MockMuxer extends AbstractStreamMuxer<MockMuxedStream> {
       log: this.log.newScope(`stream:${direction}:${id}`),
       sendMessage: this.sendMessage,
       encoding: this.encoding,
-      maxMessageSize: this.maxMessageSize - MESSAGE_OVERHEAD,
-      protocol: ''
+      maxMessageSize: this.maxMessageSize - MESSAGE_OVERHEAD
     })
   }
 }
